+++ CHANGELOG.md
@@ -12,7 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
-- Write client ID to log every time an access token is generated.
+- Write client name to log every time an access token is generated.
+- Always verify client ID if provided in token request.
 
 ## [4.13.1] - 2025-11-19